requirements.txt
django-autoslug==1.9.3
Pillow==5.3.0
psycopg2==2.7.5
pytz==2018.5
